requirements.txt
markdown
streamlit
pandas
numpy
crewai
crewai-tools
streamlit-authenticator
yfinance
pyaml-env
google-cloud-firestore
itables
streamlit-extras